requirements.txt
flask
gunicorn  # Servidor de producción para Flask
python-magic-bin  # Para validar tipos de archivos
werkzeug  # Ya incluido en Flask, pero por si acaso
